+++ R/log_parser.R
@@ -3,13 +3,18 @@
 # Take log and return processed results
 plexos_log_parser <- function(txt) {
   out <- list()
-  try(out$log_info  <- log_info(txt),  silent = !getOption("rplexos.debug"))
-  try(out$log_steps <- log_steps(txt), silent = !getOption("rplexos.debug"))
-  
-  if (!"log_info" %in% names(out))
-    rplexos_message("Log parsing failed when creating summary for each step (log_info)")
-  if (!"log_steps" %in% names(out))
-    rplexos_message("Log parsing failed when reading the step solution times (log_steps)")
+  if (length(txt) == 1L) {
+    try(out$log_info  <- log_info(txt),  silent = !is_debug_rplexos())
+    try(out$log_steps <- log_steps(txt), silent = !is_debug_rplexos())
+    
+    if (!"log_info" %in% names(out))
+      rplexos_message("Log parsing failed when creating summary for each step (log_info)")
+    if (!"log_steps" %in% names(out))
+      rplexos_message("Log parsing failed when reading the step solution times (log_steps)")
+  } else {
+    # TODO: Parse log files that are read in multiple pieces
+    rplexos_message("Log file not parsed because it is large")
+  }
   
   out
 }