+++ src/process_xml.cpp
@@ -37,16 +37,24 @@ bool column_is_id(std::string col_name) {
 }
 
 // [[Rcpp::export]]
-Rcpp::List process_xml(std::string xml) {
+Rcpp::List process_xml(Rcpp::CharacterVector xml) {
     // Start XML variables
     xml_document<> doc;
-	xml_node<> *root_node;
+    xml_node<> *root_node;
+    
+    // Merge all the strings into one
+    long xml_len = 0;
+    for (int i = 0; i < xml.size(); ++i)
+        xml_len += xml[i].size();
+    
+    vector<char> contents(xml_len + 1);
+    vector<char>::iterator it = contents.begin();
+    for (int i = 0; i < xml.size(); ++i)
+        it = copy(xml[i].begin(), xml[i].end(), it);
     
     // Parse XML file and find root node
-    vector<char> contents(xml.size() + 1);
-    copy(xml.begin(), xml.end(), contents.begin());
-	doc.parse<0> (&contents[0]);
-	root_node = doc.first_node();
+    doc.parse<0> (&contents[0]);
+    root_node = doc.first_node();
     
     // Variables for the loop
     vector<string> table_names;
@@ -56,7 +64,7 @@ Rcpp::List process_xml(std::string xml) {
     int prev_pos = -1, pos = -1;
     
     // Get list of tables and headers
-	for (xml_node<> *param_node = root_node->first_node(); param_node; param_node = param_node->next_sibling()) {
+    for (xml_node<> *param_node = root_node->first_node(); param_node; param_node = param_node->next_sibling()) {
         vector<string> names(0);
         curr_table = string(param_node->name());